Add file upload control with preview
This commit modifies the Profile component. A file upload
control and preview element for the selected file are rendered.

When user selects a file, a preview is shown.

When user clicks submit, a stub function is called a FormData
object holding the file.

Code should be added to imagesAPI.uploadImage to connect to
an image hosting service.